src/main/java/ru/clevertec/check/util/PropertiesUtil.java
package ru.clevertec.check.util;

import lombok.Getter;
import lombok.SneakyThrows;
import lombok.experimental.UtilityClass;

import java.io.IOException;
import java.util.Properties;
@UtilityClass
public final class PropertiesUtil {

    @Getter
    private static final Properties PROPERTIES = new Properties();
    @Getter
    private static final String INIT = "START";


    static {
        loadProperties();
    }

    private static void loadProperties() {
        try (var inputStream = PropertiesUtil.class.getClassLoader().getResourceAsStream("application.properties")) {
            PROPERTIES.load(inputStream);
        } catch (IOException e) {
            throw new RuntimeException(e);
        }
    }

    public static String get(String key) {
        return PROPERTIES.getProperty(key);
    }
}